readWebcamFlask/requirements.txt
flask==2.3.3
opencv-python==4.8.0
pytesseract==0.3.10
Pillow==10.0.0
numpy==1.24.3
eventlet==0.33.3  # For WebSocket support
flask-socketio==5.3.6  # For real-time updates 